requirements.txt
Flask==2.0.1
Flask-RESTful==0.3.9
Flask-JWT-Extended==4.3.1
pymongo==4.0.1
requests==2.26.0
beautifulsoup4==4.10.0
pandas==1.3.3
numpy==1.21.2
scikit-learn==1.0
pytest==6.2.5
gunicorn==20.1.0
python-dotenv==0.19.1
schedule==1.1.0
aiohttp==3.7.4
redis==3.5.3
celery==5.0.5